# docs/api/openapi.locked.yaml
openapi: 3.1.0
info:
  title: FaultMaven API
  description: AI-powered troubleshooting assistant for Engineers, SREs, and DevOps
    professionals
  version: 1.0.0
paths:
...
  /api/v1/auth/dev-login:
    post:
      tags:
      - authentication
      - authentication
      summary: Dev Login
      description: "Developer login mock endpoint.\n\nCreates or authenticates a user\
        \ with minimal validation (username/email only).\nReturns complete ViewState\
        \ for immediate UI rendering.\n\nArgs:\n    request: DevLoginRequest with\
        \ username (email)\n    session_service: Injected session service\n    \n\
        Returns:\n    AuthResponse with ViewState containing user context and initial\
        \ data"
      operationId: dev_login_api_v1_auth_dev_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/auth/session/{session_id}:
    get:
      tags:
      - authentication
      - authentication
      summary: Verify Session
      description: "Verify existing session and return current ViewState.\n\nUsed\
        \ by frontend to restore session on app startup.\n\nArgs:\n    session_id:\
        \ Session ID to verify\n    session_service: Injected session service\n  \
        \  \nReturns:\n    AuthResponse with current ViewState if session is valid"
      operationId: verify_session_api_v1_auth_session__session_id__get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AuthResponse:
      properties:
        schema_version:
          type: string
          enum:
          - 3.1.0
          const: 3.1.0
          title: Schema Version
          default: 3.1.0
        success:
          type: boolean
          title: Success
          default: true
        view_state:
          $ref: '#/components/schemas/ViewState'
      type: object
      required:
      - view_state
      title: AuthResponse
      description: Response payload for authentication operations with ViewState.
...
    Case:
      properties:
        case_id:
          type: string
          title: Case Id
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        status:
          type: string
          enum:
          - active
          - resolved
          - archived
          title: Status
          default: active
        priority:
          type: string
          enum:
          - low
          - medium
          - high
          - critical
          title: Priority
          default: medium
        created_at:
          type: string
          title: Created At
        updated_at:
          type: string
          title: Updated At
        message_count:
          type: integer
          title: Message Count
          default: 0
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      type: object
      required:
      - case_id
      - title
      title: Case
      description: Represents a troubleshooting case.
...
    DataType:
      type: string
      enum:
      - log_file
      - config_file
      - error_report
      - documentation
      - screenshot
      - other
      title: DataType
      description: Defines the type of data uploaded by users.
    DevLoginRequest:
      properties:
        username:
          type: string
          title: Username
      type: object
      required:
      - username
      title: DevLoginRequest
      description: Request payload for developer login.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProcessingStatus:
      type: string
      enum:
      - pending
      - processing
      - completed
      - failed
      title: ProcessingStatus
      description: Defines the status of data processing operations.
...
    UploadedData:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        type:
          $ref: '#/components/schemas/DataType'
        size_bytes:
          type: integer
          title: Size Bytes
        upload_timestamp:
          type: string
          title: Upload Timestamp
        processing_status:
          $ref: '#/components/schemas/ProcessingStatus'
        processing_summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Processing Summary
        confidence_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Confidence Score
      type: object
      required:
      - id
      - name
      - type
      - size_bytes
      - upload_timestamp
      - processing_status
      title: UploadedData
      description: A strongly-typed model for data uploaded by the user.
    User:
      properties:
        user_id:
          type: string
          title: User Id
        email:
          type: string
          title: Email
        name:
          type: string
          title: Name
        created_at:
          type: string
          title: Created At
        last_login:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Login
      type: object
      required:
      - user_id
      - email
      - name
      title: User
      description: Represents a user in the system.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ViewState:
      properties:
        session_id:
          type: string
          title: Session Id
        user:
          $ref: '#/components/schemas/User'
        active_case:
          anyOf:
          - $ref: '#/components/schemas/Case'
          - type: 'null'
        cases:
          items:
            $ref: '#/components/schemas/Case'
          type: array
          title: Cases
        messages:
          items:
            type: object
          type: array
          title: Messages
        uploaded_data:
          items:
            $ref: '#/components/schemas/UploadedData'
          type: array
          title: Uploaded Data
        show_case_selector:
          type: boolean
          title: Show Case Selector
          default: true
        show_data_upload:
          type: boolean
          title: Show Data Upload
          default: true
        loading_state:
          anyOf:
          - type: string
          - type: 'null'
          title: Loading State
      type: object
      required:
      - session_id
      - user
      title: ViewState
      description: 'Comprehensive view state representing the complete frontend rendering
        state.

        This is the single source of truth for what the frontend should display.'